python/requirements.txt
falcon==3.1.1
requests==2.31.0
SQLAlchemy==2.0.29
psycopg[binary]==3.1.18
python-dateutil==2.9.0.post0
google-cloud-secret-manager==2.20.0
waitress==2.1.2
python-dotenv==1.0.1
